+++ univeq/shared/src/main/scala/japgolly/univeq/UnivEq.scala
@@ -1,7 +1,9 @@
 package japgolly.univeq
 
 import java.{lang => jl}
-import scala.collection.immutable.ListSet
+import java.{time => jt}
+import scala.collection.{immutable => sci}
+import scala.concurrent.{duration => sd}
 
 /**
  * Universal equality.
@@ -19,35 +21,20 @@ object UnivEq {
   final val UnivEqAnyRef: UnivEq[AnyRef] =
     new UnivEq[AnyRef] {}
 
-  @inline def force[A]: UnivEq[A] =
-    // TODO Why does this work?!
-    // UnivEq is specialised. Casting shouldn't work for T <: AnyVal but it does...
+  def force[A]: UnivEq[A] =
     UnivEqAnyRef.asInstanceOf[UnivEq[A]]
 
-  // Primitives
-  implicit val UnivEqUnit   : UnivEq[Unit   ] = new UnivEq[Unit   ] {}
-  implicit val UnivEqChar   : UnivEq[Char   ] = new UnivEq[Char   ] {}
-  implicit val UnivEqLong   : UnivEq[Long   ] = new UnivEq[Long   ] {}
-  implicit val UnivEqInt    : UnivEq[Int    ] = new UnivEq[Int    ] {}
-  implicit val UnivEqShort  : UnivEq[Short  ] = new UnivEq[Short  ] {}
-  implicit val UnivEqBoolean: UnivEq[Boolean] = new UnivEq[Boolean] {}
-  implicit val UnivEqByte   : UnivEq[Byte   ] = new UnivEq[Byte   ] {}
-
-  // Scala
-          implicit val univEqString                         : UnivEq[String      ] = force
-  @inline implicit def univEqClass  [A]                     : UnivEq[Class    [A]] = force
-  @inline implicit def univEqClass_                         : UnivEq[Class    [_]] = force
-  @inline implicit def univEqOption [A: UnivEq]             : UnivEq[Option   [A]] = force
-  @inline implicit def univEqSet    [A: UnivEq]             : UnivEq[Set      [A]] = force
-  @inline implicit def univEqList   [A: UnivEq]             : UnivEq[List     [A]] = force
-  @inline implicit def univEqListSet[A: UnivEq]             : UnivEq[ListSet  [A]] = force
-  @inline implicit def univEqStream [A: UnivEq]             : UnivEq[Stream   [A]] = force
-  @inline implicit def univEqVector [A: UnivEq]             : UnivEq[Vector   [A]] = force
-  @inline implicit def univEqEither [A: UnivEq, B: UnivEq]  : UnivEq[Either[A, B]] = force
-  @inline implicit def univEqEitherL[A: UnivEq, B        ]  : UnivEq[Left  [A, B]] = force
-  @inline implicit def univEqEitherR[A        , B: UnivEq]  : UnivEq[Right [A, B]] = force
-  @inline implicit def univEqMap    [K: UnivEq, V: UnivEq]  : UnivEq[Map   [K, V]] = force
-  @inline implicit def univEqEnum   [A <: Enumeration#Value]: UnivEq[A]            = force
+  // Primitives & core
+  implicit val univEqUnit   : UnivEq[Unit   ] = new UnivEq[Unit   ]{}
+  implicit val univEqChar   : UnivEq[Char   ] = new UnivEq[Char   ]{}
+  implicit val univEqShort  : UnivEq[Short  ] = new UnivEq[Short  ]{}
+  implicit val univEqByte   : UnivEq[Byte   ] = new UnivEq[Byte   ]{}
+  implicit val univEqLong   : UnivEq[Long   ] = new UnivEq[Long   ]{}
+  implicit val univEqInt    : UnivEq[Int    ] = new UnivEq[Int    ]{}
+  implicit def univEqFloat  : UnivEq[Float  ] = new UnivEq[Float  ]{}
+  implicit val univEqDouble : UnivEq[Double ] = new UnivEq[Double ]{}
+  implicit val univEqBoolean: UnivEq[Boolean] = new UnivEq[Boolean]{}
+  @inline implicit def univEqString: UnivEq[String] = force
 
   // Tuples
   @inline implicit def univEqTuple2[A:UnivEq, B:UnivEq]: UnivEq[(A,B)] = force
@@ -72,13 +59,67 @@ object UnivEq {
   @inline implicit def univEqTuple21[A:UnivEq, B:UnivEq, C:UnivEq, D:UnivEq, E:UnivEq, F:UnivEq, G:UnivEq, H:UnivEq, I:UnivEq, J:UnivEq, K:UnivEq, L:UnivEq, M:UnivEq, N:UnivEq, O:UnivEq, P:UnivEq, Q:UnivEq, R:UnivEq, S:UnivEq, T:UnivEq, U:UnivEq]: UnivEq[(A,B,C,D,E,F,G,H,I,J,K,L,M,N,O,P,Q,R,S,T,U)] = force
   @inline implicit def univEqTuple22[A:UnivEq, B:UnivEq, C:UnivEq, D:UnivEq, E:UnivEq, F:UnivEq, G:UnivEq, H:UnivEq, I:UnivEq, J:UnivEq, K:UnivEq, L:UnivEq, M:UnivEq, N:UnivEq, O:UnivEq, P:UnivEq, Q:UnivEq, R:UnivEq, S:UnivEq, T:UnivEq, U:UnivEq, V:UnivEq]: UnivEq[(A,B,C,D,E,F,G,H,I,J,K,L,M,N,O,P,Q,R,S,T,U,V)] = force
 
+  // Scala Predef
+  @inline implicit def univEqBigDecimal                        : UnivEq[BigDecimal  ] = force
+  @inline implicit def univEqBigInt                            : UnivEq[BigInt      ] = force
+  @inline implicit def univEqEither    [A: UnivEq, B: UnivEq]  : UnivEq[Either[A, B]] = force
+  @inline implicit def univEqEitherL   [A: UnivEq, B        ]  : UnivEq[Left  [A, B]] = force
+  @inline implicit def univEqEitherR   [A        , B: UnivEq]  : UnivEq[Right [A, B]] = force
+  @inline implicit def univEqEnum      [A <: Enumeration#Value]: UnivEq[A]            = force
+  @inline implicit def univEqList      [A: UnivEq]             : UnivEq[List     [A]] = force
+  @inline implicit def univEqMap       [A: UnivEq, B: UnivEq]  : UnivEq[Map   [A, B]] = force
+  @inline implicit def univEqNone                              : UnivEq[None.type   ] = force
+  @inline implicit def univEqOption    [A: UnivEq]             : UnivEq[Option   [A]] = force
+  @inline implicit def univEqRange                             : UnivEq[Range       ] = force
+  @inline implicit def univEqSet       [A: UnivEq]             : UnivEq[Set      [A]] = force
+  @inline implicit def univEqSome      [A: UnivEq]             : UnivEq[Some     [A]] = force
+  @inline implicit def univEqStream    [A: UnivEq]             : UnivEq[Stream   [A]] = force
+  @inline implicit def univEqVector    [A: UnivEq]             : UnivEq[Vector   [A]] = force
+
+  // scala.collection.immutable
+  @inline implicit def univEqSciBitSet   [A: UnivEq           ]: UnivEq[sci.BitSet         ] = force
+  @inline implicit def univEqSciHashMap  [A: UnivEq, B: UnivEq]: UnivEq[sci.HashMap  [A, B]] = force
+  @inline implicit def univEqSciHashSet  [A: UnivEq           ]: UnivEq[sci.HashSet  [A]   ] = force
+  @inline implicit def univEqSciIntMap   [A: UnivEq           ]: UnivEq[sci.IntMap   [A]   ] = force
+  @inline implicit def univEqSciListMap  [A: UnivEq, B: UnivEq]: UnivEq[sci.ListMap  [A, B]] = force
+  @inline implicit def univEqSciListSet  [A: UnivEq           ]: UnivEq[sci.ListSet  [A]   ] = force
+  @inline implicit def univEqSciLongMap  [A: UnivEq           ]: UnivEq[sci.LongMap  [A]   ] = force
+  @inline implicit def univEqSciSortedMap[A: UnivEq, B: UnivEq]: UnivEq[sci.SortedMap[A, B]] = force
+  @inline implicit def univEqSciSortedSet[A: UnivEq           ]: UnivEq[sci.SortedSet[A]   ] = force
+  @inline implicit def univEqSciTreeMap  [A: UnivEq, B: UnivEq]: UnivEq[sci.TreeMap  [A, B]] = force
+  @inline implicit def univEqSciTreeSet  [A: UnivEq           ]: UnivEq[sci.TreeSet  [A]   ] = force
+
+  // scala.concurrent.duration
+  @inline implicit def univEqSDuration      : UnivEq[sd.Duration      ] = force
+  @inline implicit def univEqSFiniteDuration: UnivEq[sd.FiniteDuration] = force
+
   // Java
+  @inline implicit def univEqClass  [A]            : UnivEq[Class  [A]] = force
+  @inline implicit def univEqClass_                : UnivEq[Class  [_]] = force
+  @inline implicit def univEqJDouble               : UnivEq[jl.Double ] = force
+  @inline implicit def univEqJFloat                : UnivEq[jl.Float  ] = force
   @inline implicit def univEqJInteger              : UnivEq[jl.Integer] = force
   @inline implicit def univEqJLong                 : UnivEq[jl.Long   ] = force
   @inline implicit def univEqJBoolean              : UnivEq[jl.Boolean] = force
   @inline implicit def univEqJByte                 : UnivEq[jl.Byte   ] = force
   @inline implicit def univEqJShort                : UnivEq[jl.Short  ] = force
   @inline implicit def univEqJEnum[A <: jl.Enum[A]]: UnivEq[A         ] = force
+
+  // java.time
+  @inline implicit def univEqJavaTimeDuration      : UnivEq[jt.Duration      ] = force
+  @inline implicit def univEqJavaTimeInstant       : UnivEq[jt.Instant       ] = force
+  @inline implicit def univEqJavaTimeLocalDate     : UnivEq[jt.LocalDate     ] = force
+  @inline implicit def univEqJavaTimeLocalDateTime : UnivEq[jt.LocalDateTime ] = force
+  @inline implicit def univEqJavaTimeLocalTime     : UnivEq[jt.LocalTime     ] = force
+  @inline implicit def univEqJavaTimeMonthDay      : UnivEq[jt.MonthDay      ] = force
+  @inline implicit def univEqJavaTimeOffsetDateTime: UnivEq[jt.OffsetDateTime] = force
+  @inline implicit def univEqJavaTimeOffsetTime    : UnivEq[jt.OffsetTime    ] = force
+  @inline implicit def univEqJavaTimePeriod        : UnivEq[jt.Period        ] = force
+  @inline implicit def univEqJavaTimeYear          : UnivEq[jt.Year          ] = force
+  @inline implicit def univEqJavaTimeYearMonth     : UnivEq[jt.YearMonth     ] = force
+  @inline implicit def univEqJavaTimeZonedDateTime : UnivEq[jt.ZonedDateTime ] = force
+  @inline implicit def univEqJavaTimeZoneId        : UnivEq[jt.ZoneId        ] = force
+  @inline implicit def univEqJavaTimeZoneOffset    : UnivEq[jt.ZoneOffset    ] = force
 
   // Derivation
   @inline def derive            [A <: AnyRef]: UnivEq[A] = macro macros.UnivEqMacros.deriveAutoQuiet[A]